+++ tests/Fixtures/Mesh/Forms/Input.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Mesh\Forms;
+
+use EthanBarlo\Mesh\Component;
+
+class Input extends Component {}